Added file_exists method.
